HUB-971: Update tests to use the new "about documents" page
The "select documents" page has been replaced with the new page. In order for the IDP picker to work, we still need to click the button on the static "about documents" page as that fakes the document selections. This will be removed in a future frontend PR and the tests will have to be updated again.

Additional improvements:

  - Add a script to easily run tests against different environments

  - Make indentation the same across all files. Some files in the project were indented with 2 spaces and some with 4. This commit makes all files use 4 spaces for consistency.

  - Use 'let' instead of 'var' in all places